requirements.txt
Flask==2.2.2
Werkzeug==2.3.7 # for flask + python3.8^
requests==2.28.1
PyJWT==2.7.0
Flask-JWT-Extended==4.5.2
PyYAML==6.0.1
Flask-Cors == 4.0.0